lectures/Web-based System (2024-2)/team-G/requirements.txt
Flask
shap
pandas
numpy
mne
torch
Pillow
matplotlib
scikit-learn